requirements.txt
# Django
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0

# Authentication
djangorestframework-simplejwt>=5.0,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Task Queue
celery>=5.2,<6.0
django-celery-beat>=2.0,<3.0
django-celery-results>=2.0,<3.0
kombu>=5.0,<6.0

# Cache
redis>=5.0,<6.0
django-redis>=6.0,<7.0

# WebSockets
channels>=4.0,<5.0
channels-redis>=4.0,<5.0

# API Tools
drf-yasg>=1.20,<2.0

# Email
python-dotenv>=1.0,<2.0

# Development
django-extensions>=4.0,<5.0

# Production Server
gunicorn>=21.0,<24.0
gevent>=24.0,<25.0
uvicorn>=0.30,<1.0

# Payment Providers
stripe>=7.0,<8.0
paypalrestsdk>=1.13,<2.0

# Notification Providers
sendgrid>=6.0,<7.0
twilio>=8.0,<9.0
firebase-admin>=6.0,<7.0

# Utilities
requests>=2.32,<3.0
pytz>=2024,<2025

# Structured Logging
structlog>=24.0,<25.0
python-json-logger>=2.0,<3.0
colorama>=0.4.6
